feat: cacheable SSR + ISR for /, /products, and /products/[id] (#6)
- Add `revalidate = 60` to all three primary routes so rendered HTML is
  cached via the Harper cache handler and regenerated on a TTL.
- Convert /products to a server component; interactive filter/sort UI moves
  to a new products-browser.js client component seeded with server-fetched
  products, so the product grid is in the initial HTML.
- Remove generateStaticParams from the PDP (no more build-time product
  enumeration/database lock); unknown IDs render on demand and a server-side
  notFound() guard returns 404 instead of a serialization crash.
- Add per-route Cache-Control (public, max-age=60, stale-while-revalidate=600)
  via headers() in next.config.js.
- Unit tests cover the headers() contract and route-segment structure.
