+++ next.config.js
@@ -9,6 +9,16 @@ const nextConfig = {
 	images: { unoptimized: true },
 	cacheHandler: CACHE_HANDLER_PATH,
 	cacheMaxMemorySize: 0,
+	// Deterministic per-route Cache-Control for the cacheable SSR/ISR routes
+	// (issue #6). Values align with the routes' `revalidate = 60`.
+	async headers() {
+		const cacheControl = { key: 'Cache-Control', value: 'public, max-age=60, stale-while-revalidate=600' };
+		return [
+			{ source: '/', headers: [cacheControl] },
+			{ source: '/products', headers: [cacheControl] },
+			{ source: '/products/:id', headers: [cacheControl] },
+		];
+	},
 };
 
 module.exports = withHarper(nextConfig);
